added features and mappers for the miam domestic voilence screen
