+++ src/main/steps/c100-rebuild/check-your-answers/content.ts
@@ -187,6 +187,12 @@ export const enContent = {
     financiallyAbuse_hint:
       'Financial abuse is a way of controlling someone being able to earn, spend or keep their own money. For example, preventing someone going to work, withholding money, or putting debts in someone else’s name. Evidence could include: <ul class="govuk-list govuk-list--bullet govuk-hint govuk-!-margin-top-2"> <li>a copy of a credit card account, loan document or bank statements</li><li>a letter from a domestic violence support organisation</li><li>emails, text messages or a diary kept by the victim</li> </ul>',
     noneOfOptions: 'None of the above',
+    localAuthority:
+      'The children in the application, or another child in the household, is the subject of enquiries by a local authority under section 47 of the Children Act 1989 Act',
+    localAuthorityHint:
+      'This may mean that a local authority is carrying out enquiries because of concerns the children are suffering or might suffer significant harm. See <a href="https://www.legislation.gov.uk/ukpga/1989/41/section/17" class="govuk-link" target="_blank" aria-label="section 47 of the Children Act 1989 Act">section 47 of the Children Act 1989 Act</a> for further details.',
+    childProtectionPlan:
+      'The children in the application, or another child in the household, is the subject of a child protection plan put in place by the local authority',
   },
 };
 const cyContent: typeof enContent = {
@@ -361,6 +367,12 @@ const cyContent: typeof enContent = {
     financiallyAbuse_hint:
       'Financial abuse is a way of controlling someone being able to earn, spend or keep their own money. For example, preventing someone going to work, withholding money, or putting debts in someone else’s name. Evidence could include: <ul class="govuk-list govuk-list--bullet govuk-hint govuk-!-margin-top-2"> <li>a copy of a credit card account, loan document or bank statements</li><li>a letter from a domestic violence support organisation</li><li>emails, text messages or a diary kept by the victim</li> </ul>',
     noneOfOptions: 'None of the above',
+    localAuthority:
+      'The children in the application, or another child in the household, is the subject of enquiries by a local authority under section 47 of the Children Act 1989 Act',
+    localAuthorityHint:
+      'This may mean that a local authority is carrying out enquiries because of concerns the children are suffering or might suffer significant harm. See <a href="https://www.legislation.gov.uk/ukpga/1989/41/section/17" class="govuk-link" target="_blank" aria-label="section 47 of the Children Act 1989 Act">section 47 of the Children Act 1989 Act</a> for further details.',
+    childProtectionPlan:
+      'The children in the application, or another child in the household, is the subject of a child protection plan put in place by the local authority',
   },
 };
